+++ src/components/Profile.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import TopicCard from './TopicCard';
 import LinkButton from './LinkButton';
 import { Candidate, SideJob } from '../Types';
-import VoteCard from '../components/VoteCard';
+import VoteCard from '../components/VoteCard/VoteCard';
 import SideJobCard from '../components/SideJobCard';
 import './Profile.css';
 import TitleHeader from './TitleHeader';
