+++ src/components/VoteCard/VoteCard.css
