+++ src/components/VoteCard/VoteChart/VoteChart.css
