+++ src/components/VoteCard/VoteChart/VoteChart.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import './VoteChart.css';
 import { HorizontalBar } from 'react-chartjs-2';
-import { Poll } from '../Types';
-import '../index.css';
+import { Poll } from '../../../Types';
+import '../../../index.css';
 
 /* 
 Register the ChartProps interface so we can pass properties to the component.
