+++ src/components/VoteCard/VoteDetails/InfoText/InfoText.css
