+++ src/components/VoteCard/VoteDetails/InfoText/Infotext.tsx
