+++ src/components/VoteCard/VoteDetails/Middlebar/Middlebar.css
