+++ src/components/VoteCard/VoteDetails/Middlebar/Middlebar.tsx
