+++ src/components/VoteCard/VoteDetails/Preview/Preview.css
