+++ src/components/VoteCard/VoteDetails/Preview/Preview.tsx
