+++ src/components/VoteCard/VoteDetails/Table/Content/Columns/Column/Column.css
