+++ src/components/VoteCard/VoteDetails/Table/Content/Columns/Columns.tsx
