+++ src/components/VoteCard/VoteDetails/Table/Content/Columns/TotalColumn/TotalColumn.tsx
