+++ src/components/VoteCard/VoteDetails/Table/Content/Content.css
