+++ src/components/VoteCard/VoteDetails/Table/NumberChips/NumberChip/NumberChip.tsx
