+++ src/components/VoteCard/VoteDetails/Table/NumberChips/NumberChips.css
