+++ src/components/VoteCard/VoteDetails/Table/NumberChips/NumberChips.tsx
