+++ src/components/VoteCard/VoteDetails/Table/PartyChips/PartyChip/PartyChip.css
