+++ src/components/VoteCard/VoteDetails/Table/PositioningColumns/PositioningColumns.tsx
