+++ src/components/VoteCard/VoteDetails/Table/Table.css
