+++ src/components/VoteCard/VoteDetails/Table/Table.tsx
