+++ src/components/VoteCard/VoteDetails/VoteDetails.tsx
