+++ src/components/VoteCard/VoteDetails/VotesDetailsToolbar/VoteDetailsToolbar.tsx
