+++ src/pages/Votes.tsx
@@ -1,6 +1,6 @@
 import { IonContent, IonPage } from '@ionic/react';
 import React from 'react';
-import VoteCard from '../components/VoteCard';
+import VoteCard from '../components/VoteCard/VoteCard';
 import './Votes.css';
 import { Candidate } from '../Types';
 import LinkButton from '../components/LinkButton';
